fix: Tabbar works & DBClass
Made it so Tabbar is not visible in login & Register, only after logging in or Registering you are able to see the tabbar. Also Changed the connectURI so now everyone can use the link due to it being IP locked in the passed. Deleted unused files, Also put Auth0 spotify in comments to prevent errors

Reviewed-by: Axel
Version: 1